% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{data_analyze}
\alias{data_analyze}
\title{Function to produce analysis of the data and print the results.}
\usage{
data_analyze(datapath = NULL, savedir = "./")
}
\arguments{
\item{datapath}{is the where the NC Court and Demography data are saved.}

\item{savedir}{is the folder where the analysis should be saved.}
}
\description{
Function to produce analysis of the data and print the results.
}
