% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{data_download}
\alias{data_download}
\title{Function to download data from the NC Judicial Branch and the Office of
State Budget and Management}
\usage{
data_download(filepath = "./data.rds.xz")
}
\arguments{
\item{filepath}{is where the download should be saved. the default is the
current directory.}
}
\description{
Function to download data from the NC Judicial Branch and the Office of
State Budget and Management
}
