% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{data_process}
\alias{data_process}
\title{Function to process the data and store a cleaned version on disk}
\usage{
data_process(data)
}
\arguments{
\item{datapath}{is the where the NC Court and Demography data are saved.}
}
\description{
Function to process the data and store a cleaned version on disk
}
